Move the `api-server` module under the `@bayou` namespace.
