+++ local-modules/@bayou/api-server/tests/test_BearerToken.js
@@ -5,11 +5,11 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { BearerToken } from 'api-server';
+import { BearerToken } from '@bayou/api-server';
 
 const SECRET_TOKEN = 'Setec-Astronomy-Setec-Astronomy-';
 
-describe('api-server/BearerToken', () => {
+describe('@bayou/api-server/BearerToken', () => {
   describe('constructor(secret)', () => {
     it('should reject secrets with length < 32', () => {
       assert.throws(() => new BearerToken('Setec-Astronomy'));
