+++ local-modules/@bayou/hooks-server/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
 
   "dependencies": {
-    "api-server": "local",
+    "@bayou/api-server": "local",
     "file-store-local": "local",
     "@bayou/hooks-common": "local",
     "@bayou/promise-util": "local",
